Capstone1_Project/requirements.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
xgboost
joblib
pyyaml
scipy
fastapi
uvicorn[standard]
notebook
ipykernel
